+++ apps/web/src/models/History.ts
@@ -1,0 +1,12 @@
+import { Prize } from "./Campaign"
+import { Customer } from "./Customer"
+
+
+export type History = {
+  id: string
+  date: string
+  status: 'claimed' | 'unclaimed'
+  prize: Prize
+  customer: Customer
+}
+