+++ packages/components/src/composite/index.ts
@@ -1,7 +1,9 @@
 
+export * from './data-table'
 export * from './file-upload'
 export * from './icon'
 export * from './multiple-selector'
 export * from './slideover'
 export * from './popup'
 export * from './creatable-combobox'
+export * from './qr-code'